fixed first creation error
